+++ packages/cli-core/GitUtilities.js
@@ -1,5 +1,6 @@
-const execa = require('execa');
 const path = require('path');
+
+const execa = require('execa');
 const slash = require('slash');
 
 function hasTag(tag) {
@@ -13,45 +14,45 @@ function hasTag(tag) {
     () => false,
   );
 }
-const repoName = name => name.replace(/^@.+\//, '');
+const repoName = (name) => name.replace(/^@.+\//, '');
 
 exports.repoName = repoName;
 
-exports.getRemoteUrl = cwd =>
+exports.getRemoteUrl = (cwd) =>
   execa('git', ['config', 'remote.origin.url'], { cwd })
-    .then(r => r.stdout)
+    .then((r) => r.stdout)
     .catch(() => '');
 
 exports.remoteUrl = (name, org = '4Catalyzer') =>
   `https://github.com/${org}/${repoName(name)}.git`;
 
-exports.isGitRepo = dest =>
+exports.isGitRepo = (dest) =>
   execa('git', ['rev-parse'], { cwd: dest, stdio: 'inherit' })
     .then(() => true)
     .catch(() => false);
 
-exports.init = dest =>
+exports.init = (dest) =>
   exports
     .isGitRepo(dest)
     .then(
-      isGit =>
+      (isGit) =>
         !isGit && execa('git', ['init'], { cwd: dest, stdio: 'inherit' }),
     );
 
-exports.addFile = file =>
+exports.addFile = (file) =>
   execa('git', [
     'add',
     slash(path.relative(process.cwd(), path.resolve(process.cwd(), file))),
   ]);
 
-exports.commit = message =>
+exports.commit = (message) =>
   execa('git', ['commit', '--no-verify', '-m', message]);
 
 exports.removeLastCommit = () => execa('git', ['reset', 'HEAD~1', '--hard']);
 
 exports.assertClean = async () => {
   const result = await execa('git', ['status', '--porcelain']).then(
-    d => d.stdout,
+    (d) => d.stdout,
   );
 
   if (result !== '')
@@ -66,7 +67,7 @@ exports.assertMatchesRemote = async () => {
     '--count',
     '--left-only',
     '@{u}...HEAD',
-  ]).then(d => d.stdout);
+  ]).then((d) => d.stdout);
 
   if (result !== '0')
     throw new Error('The remote differs from the local working tree');
@@ -76,20 +77,20 @@ exports.addRemote = (dest, name, org) =>
   execa('git', ['remote', 'add', 'origin', exports.remoteUrl(name, org)], {
     cwd: dest,
     stdio: 'inherit',
-  }).catch(err => {
+  }).catch((err) => {
     if ((err.stderr || '').match(/remote origin already exists/)) return;
     throw err;
   });
 
-exports.addTag = async tag => {
+exports.addTag = async (tag) => {
   if (await hasTag(tag)) return;
   await execa('git', ['tag', tag, '-m', tag]);
 };
 
-exports.removeTag = tag => execa('git', ['tag', '-d', tag]);
-exports.getCurrentBranch = opts =>
+exports.removeTag = (tag) => execa('git', ['tag', '-d', tag]);
+exports.getCurrentBranch = (opts) =>
   execa('git', ['rev-parse', '--abbrev-ref', 'HEAD'], opts).then(
-    r => r.stdout,
+    (r) => r.stdout,
   );
 
 exports.pushWithTags = () => execa('git', ['push', '--follow-tags']);