+++ packages/cli/index.js
@@ -1,19 +1,22 @@
 #!/usr/bin/env node
 
 const build = require('@4c/build/command');
-const start = require('@4c/start/command');
 const init = require('@4c/init/command');
-const rollout = require('@4c/rollout/command');
-const yargs = require('yargs');
 const intl = require('@4c/intl/command');
+const rollout = require('@4c/rollout/command');
+const start = require('@4c/start/command');
 const format = require('pedantic/format');
 const lint = require('pedantic/lint');
 const svg2c = require('svg2c/command');
+const workspaces = require('ts-doctor/workspaces');
+const yargs = require('yargs');
 
-const setCmdName = (name, cmd) => ({
-  ...cmd,
-  command: cmd.command.replace(/^\$0/, name),
-});
+function setCmdName(name, cmd) {
+  return {
+    ...cmd,
+    command: cmd.command.replace(/^\$0/, name),
+  };
+}
 
 yargs
   .help()
@@ -30,6 +33,7 @@ yargs
   .command(setCmdName('release', rollout))
   .command(setCmdName('intl', intl))
   .command(setCmdName('icons', svg2c))
-  .command(format)
-  .command(lint)
+  .command(setCmdName('fixup-workspaces', workspaces))
+  .command(setCmdName('format', format))
+  .command(setCmdName('lint', lint))
   .parse(process.argv.slice(2));