+++ packages/file-butler/lib.js
@@ -1,10 +1,11 @@
 const path = require('path');
+
+const cpy = require('cpy');
 const fs = require('fs-extra');
 const globby = require('globby');
-const cpy = require('cpy');
 
 const strReg = /\[\s*(\w+)\s*\]/g;
-const interpolate = pattern => filename => {
+const interpolate = (pattern) => (filename) => {
   const extname = path.extname(filename);
   const params = {
     extname,
@@ -23,7 +24,7 @@ async function findReadme() {
     absolute: true,
     deep: false,
     case: false,
-    transform: fp => path.normalize(fp),
+    transform: (fp) => path.normalize(fp),
   });
   return readmePath;
 }
@@ -33,7 +34,7 @@ async function findLicense() {
     absolute: true,
     deep: false,
     case: false,
-    transform: fp => path.normalize(fp),
+    transform: (fp) => path.normalize(fp),
   });
   return licensePath;
 }