+++ packages/init/command.js
@@ -1,4 +1,6 @@
-const nodePlop = require('node-plop');
+// Cherry pick to avoid pulling in all of core-js
+// see: https://github.com/plopjs/node-plop/pull/163
+const nodePlop = require('node-plop/lib/node-plop').default;
 
 // load an instance of plop from a plopfile
 
@@ -6,7 +8,7 @@ exports.command = '$0 [location]';
 
 exports.describe = 'create a new package';
 
-exports.builder = _ =>
+exports.builder = (_) =>
   _.positional('location', {
     type: 'string',
     default: process.cwd(),
@@ -22,7 +24,7 @@ exports.handler = async ({ location }) => {
   const result = await newPkg.runActions(answers);
   if (result.failures) {
     result.failures.forEach(
-      f =>
+      (f) =>
         f.error &&
         !f.error.startsWith('Aborted due to previous') &&
         console.error(f.error),