+++ packages/rollout/handleNpmError.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-param-reassign */
-const listrInput = require('listr-input');
 const chalk = require('chalk');
+const listrInput = require('listr-input');
 const { throwError } = require('rxjs');
 const { catchError } = require('rxjs/operators');
 
@@ -15,12 +15,12 @@ const handleNpmError = (error, task, executor, prompt) => {
     task.title = `${title} ${chalk.yellow('(waiting for input…)')}`;
 
     return listrInput(prompt || 'Enter OTP:', {
-      done: otp => {
+      done: (otp) => {
         task.title = title;
         return executor(otp);
       },
     }).pipe(
-      catchError(err =>
+      catchError((err) =>
         handleNpmError(err, task, executor, 'OTP was incorrect, try again:'),
       ),
     );