+++ packages/start/formatters/index.js
@@ -1,9 +1,9 @@
+const defaultError = require('./defaultError');
 const moduleNotFound = require('./moduleNotFound');
 const typescript = require('./typescript');
 const unusedFiles = require('./unusedFiles');
-const defaultError = require('./defaultError');
 
-module.exports = compiler => [
+module.exports = (compiler) => [
   unusedFiles,
   typescript(compiler),
   moduleNotFound,