+++ packages/intl/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.17](https://github.com/4Catalyzer/cli/compare/@4c/intl@1.2.16...@4c/intl@1.2.17) (2020-05-04)
+
+**Note:** Version bump only for package @4c/intl
+
+
+
+
+
+## [1.2.16](https://github.com/4Catalyzer/cli/compare/@4c/intl@1.2.15...@4c/intl@1.2.16) (2020-04-29)
+
+**Note:** Version bump only for package @4c/intl
+
+
+
+
+
+## [1.2.15](https://github.com/4Catalyzer/cli/compare/@4c/intl@1.2.14...@4c/intl@1.2.15) (2020-04-09)
+
+**Note:** Version bump only for package @4c/intl
+
+
+
+
+
+## [1.2.14](https://github.com/4Catalyzer/cli/compare/@4c/intl@1.2.13...@4c/intl@1.2.14) (2020-03-30)
+
+**Note:** Version bump only for package @4c/intl
+
+
+
+
+
+## [1.2.13](https://github.com/4Catalyzer/cli/compare/@4c/intl@1.2.12...@4c/intl@1.2.13) (2020-03-22)
+
+**Note:** Version bump only for package @4c/intl
+
+
+
+
+
+## [1.2.12](https://github.com/4Catalyzer/cli/compare/@4c/intl@1.2.11...@4c/intl@1.2.12) (2020-03-12)
+
+**Note:** Version bump only for package @4c/intl
+
+
+
+
+
+## [1.2.11](https://github.com/4Catalyzer/cli/compare/@4c/intl@1.2.10...@4c/intl@1.2.11) (2020-02-13)
+
+**Note:** Version bump only for package @4c/intl
+
+
+
+
+
 ## [1.2.10](https://github.com/4Catalyzer/cli/compare/@4c/intl@1.2.9...@4c/intl@1.2.10) (2019-11-25)
 
 **Note:** Version bump only for package @4c/intl