+++ packages/rollout/CHANGELOG.md
@@ -3,6 +3,83 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.8](https://github.com/4Catalyzer/cli/compare/@4c/rollout@2.1.7...@4c/rollout@2.1.8) (2020-05-04)
+
+**Note:** Version bump only for package @4c/rollout
+
+
+
+
+
+## [2.1.7](https://github.com/4Catalyzer/cli/compare/@4c/rollout@2.1.6...@4c/rollout@2.1.7) (2020-04-29)
+
+**Note:** Version bump only for package @4c/rollout
+
+
+
+
+
+## [2.1.6](https://github.com/4Catalyzer/cli/compare/@4c/rollout@2.1.5...@4c/rollout@2.1.6) (2020-04-09)
+
+
+### Bug Fixes
+
+* preid behavior ([#194](https://github.com/4Catalyzer/cli/issues/194)) ([766bb15](https://github.com/4Catalyzer/cli/commit/766bb15509af514e117c5af5915cee2cb25ecfc6))
+
+
+
+
+
+## [2.1.5](https://github.com/4Catalyzer/cli/compare/@4c/rollout@2.1.4...@4c/rollout@2.1.5) (2020-03-30)
+
+
+### Bug Fixes
+
+* Don't normalize package.json ([#187](https://github.com/4Catalyzer/cli/issues/187)) ([63bdd20](https://github.com/4Catalyzer/cli/commit/63bdd2002f59b1014dfd2f8db522298160bb1bfb))
+* Fail on not making commit ([#188](https://github.com/4Catalyzer/cli/issues/188)) ([e97d9c3](https://github.com/4Catalyzer/cli/commit/e97d9c37a497b8641be7a888ffb416a13f09824b))
+
+
+
+
+
+## [2.1.4](https://github.com/4Catalyzer/cli/compare/@4c/rollout@2.1.3...@4c/rollout@2.1.4) (2020-03-22)
+
+
+### Bug Fixes
+
+* Clean up dependencies ([2479dd7](https://github.com/4Catalyzer/cli/commit/2479dd743fbff67cbdb6a79f70dd3bdd00518003))
+
+
+
+
+
+## [2.1.3](https://github.com/4Catalyzer/cli/compare/@4c/rollout@2.1.2...@4c/rollout@2.1.3) (2020-03-12)
+
+
+### Bug Fixes
+
+* Fix config handling ([#180](https://github.com/4Catalyzer/cli/issues/180)) ([a400d6c](https://github.com/4Catalyzer/cli/commit/a400d6ca0b3ee133a8d2d33e5c0224cb10b0c19c))
+
+
+
+
+
+## [2.1.2](https://github.com/4Catalyzer/cli/compare/@4c/rollout@2.1.1...@4c/rollout@2.1.2) (2020-02-13)
+
+**Note:** Version bump only for package @4c/rollout
+
+
+
+
+
+## [2.1.1](https://github.com/4Catalyzer/cli/compare/@4c/rollout@2.1.0...@4c/rollout@2.1.1) (2020-02-13)
+
+**Note:** Version bump only for package @4c/rollout
+
+
+
+
+
 # [2.1.0](https://github.com/4Catalyzer/cli/compare/@4c/rollout@2.0.3...@4c/rollout@2.1.0) (2019-11-25)